backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.34
Pillow==10.2.0
stripe==7.11.0
python-dotenv==1.0.1
gunicorn==21.2.0
django-storages==1.14.2
django-redis==5.4.0
django-filter==23.5
django-rest-auth==0.9.5
django-allauth==0.60.1
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0
google-generativeai==0.3.2 